fix: sample the guest hypervisor's GS only once the base is plausible
The first version latched the first sample taken with running_l2 false,
and that is long before the guest hypervisor has set up GS. It read
gs_base = 0 on every processor and answered 0x32403206 alike - guest
linear address 8, read in early boot - which is not a processor index on
any of them.

It failed the check stated for it when it was added, that on a working
processor the value must equal that processor's own index, and that is
why the check was stated. Seven instruments in this investigation have
now reported plausible nonsense; this one reported implausible nonsense,
which is strictly better.

Now gated on a canonical kernel base, and overwritten rather than latched
so the value tracks the guest hypervisor once it is actually running
there. The counter becomes how many samples were taken, which is itself
worth seeing: zero would mean the guest hypervisor never runs with a
kernel GS base on that processor.

+++ hypervisor/include/zpp/hypervisor/hypervisor.h
@@ -7613,7 +7613,7 @@ private:
      */
     std::uint64_t l1_gs_base[max_cpus]{};
     std::uint32_t l1_gs_index[max_cpus]{};
-    std::uint8_t l1_gs_index_taken[max_cpus]{};
+    std::uint64_t l1_gs_index_taken[max_cpus]{};
 
     std::uint64_t start_up_applied[max_cpus]{};
     std::uint64_t init_emulated[max_cpus]{};
